refactor task responses
