+++ clients/python-client/src/chunkr_ai/api/task_async.py
@@ -1,34 +1,23 @@
-import asyncio
-from pydantic import BaseModel, PrivateAttr
-from datetime import datetime
-from typing import Optional, Union
-from .task_base import TaskBase
-from .protocol import ChunkrClientProtocol
-from .config import Configuration, OutputResponse, Status
+from .config import Configuration
 from .misc import prepare_upload_data
+from .task_base import TaskBase
+import asyncio
 
-class TaskResponseAsync(BaseModel, TaskBase):
-    configuration: Configuration
-    created_at: datetime
-    expires_at: Optional[datetime]
-    file_name: Optional[str]
-    finished_at: Optional[datetime]
-    input_file_url: Optional[str]
-    message: str
-    output: Optional[OutputResponse]
-    page_count: Optional[int]
-    pdf_url: Optional[str]
-    started_at: Optional[datetime]
-    status: Status
-    task_id: str
-    task_url: Optional[str]
-    _client: Optional[Union[ChunkrClientProtocol]] = PrivateAttr(default=None)
-
-    def with_client(self, client: Union[ChunkrClientProtocol]) -> 'TaskResponseAsync':
-        self._client = client
-        return self
+class TaskResponseAsync(TaskBase):
+    async def _poll_request(self) -> dict:
+        try:
+            r = await self._client._client.get(self.task_url, headers=self._client._headers())
+            r.raise_for_status()
+            return r.json()
+        except (ConnectionError, TimeoutError) as _:
+            print("Connection error while polling the task, retrying...")
+            await asyncio.sleep(0.5)
+        except Exception as e:
+            raise
 
     async def poll(self) -> 'TaskResponseAsync':
+        if not self.task_url:
+            raise ValueError("Task URL not found")
         while True:
             j = await self._poll_request()
             updated = TaskResponseAsync(**j).with_client(self._client)
@@ -36,28 +25,6 @@ class TaskResponseAsync(BaseModel, TaskBase):
             if res := self._check_status():
                 return res
             await asyncio.sleep(0.5)
-
-    async def _poll_request(self) -> dict:
-        if not self.task_url:
-            raise ValueError("Task URL not found")
-        while True:
-            try:
-                r = await self._client._client.get(self.task_url, headers=self._client._headers())
-                r.raise_for_status()
-                return r.json()
-            except Exception as e:
-                if self.status == Status.FAILED:
-                    raise ValueError(self.message) from e
-                await asyncio.sleep(0.5)
-
-    def _check_status(self) -> Optional['TaskResponseAsync']:
-        if self.status == Status.FAILED:
-            raise ValueError(f"Task failed: {self.message}")
-        if self.status == Status.CANCELLED:
-            return self
-        if self.status not in [Status.STARTING, Status.PROCESSING]:
-            return self
-        return None
 
     async def update(self, config: Configuration) -> 'TaskResponseAsync':
         if not self.task_url:
@@ -77,27 +44,7 @@ class TaskResponseAsync(BaseModel, TaskBase):
         return await self.poll()
 
     async def delete(self):
+        if not self.task_url:
+            raise ValueError("Task URL not found")
         r = await self._client._client.delete(self.task_url, headers=self._client._headers())
         r.raise_for_status()
-
-    def html(self) -> str:
-        return self._get_content("html")
-
-    def markdown(self) -> str:
-        return self._get_content("markdown")
-
-    def content(self) -> str:
-        return self._get_content("content")
-
-    def _get_content(self, t: str) -> str:
-        if not self.output:
-            return ""
-        parts = []
-        for c in self.output.chunks:
-            for s in c.segments:
-                v = getattr(s, t)
-                if v:
-                    parts.append(v)
-        return "\n".join(parts)
-
-    # Satisfying TaskBase abstract methods with stubs
